DEV: moved all of Jameson's segmentation code into a single directory to get rid no non-standard directory structure
