+++ Code/Source/sv3/Segmentation/cvITKLevelSet.h
@@ -41,8 +41,8 @@
  *      In addition, it should include helper functions to convert data between the two systems.
  */
 
-#ifndef CVITKLEVELSETBASE_H_
-#define CVITKLEVELSETBASE_H_
+#ifndef CVITKLEVELSET_H_
+#define CVITKLEVELSET_H_
 
 #include "SimVascular.h"
 #include "svSegITKExports.h" // For exports
@@ -54,9 +54,13 @@
 #define cvStructuredPoints cvStrPts
 #endif
 
-#include "Include/cvITKMacros.h"
-#include "Include/cvMacros.h"
-#include "Include/ConnectVTKITK.h"
+#include <sstream>
+#include <iostream>
+#include <string>
+
+#include "cvITKMacros.h"
+#include "cvMacros.h"
+#include "ConnectVTKITK.h"
 #include "MyUtils.h"
 
 #include "itkVTKImageExport.h"
@@ -77,29 +81,27 @@
 
 #include "cvITKUtils.h"
 
-#ifndef NULL
-#define NULL   ((void *) 0)
-#endif
 
+#define CVITKException(x)																\
+		{																				\
+	std::ostringstream message;															\
+	message << "(" << this << "): " "x";												\
+	message << __FILE__ << " " << __LINE__ << " ";										\
+	message << "ERROR: " << this->GetNameOfClass() << " ";								\
+	message << " " << __PRETTY_FUNCTION__ << std::endl;									\
+	throw std::runtime_error(message.str()); 											\
+		}																				\
 
-template<typename TInputImage = itk::Image<short,2>,
-typename TInternalPixelType = float>
-class SV_EXPORT_SEGITK cvITKLevelSetBase
+class SV_EXPORT_SEGITK cvITKLevelSet
 {
 
 public:
 
 	/** Typedefs */
 	// Depending on image size, this might be better with float...
-	typedef cvITKLevelSetBase Self;
-	typedef TInputImage ITKExternalImageType;
-	typedef itk::Image<TInternalPixelType,
-			ITKExternalImageType::ImageDimension> ITKInternalImageType;
-	enum { ImageDimension = TInputImage::ImageDimension };
-
-
-	//typedef itk::Image<short,N> ITKExternalImageType;
-	//typedef itk::Image<float,N> ITKInternalImageType;
+	typedef cvITKLevelSet Self;
+	typedef itk::Image<short,2> ITKExternalImageType;
+	typedef itk::Image<float,2> ITKInternalImageType;
 
 	typedef itk::VTKImageImport<ITKExternalImageType> itkExternalImportType;
 	typedef itk::VTKImageExport<ITKExternalImageType> itkExternalExportType;
@@ -126,18 +128,17 @@ public:
 	int ComputePhaseOneLevelSet(float kc,
 			float expFactorRising,float expFactorFalling);
 	int ComputePhaseTwoLevelSet(float kupp,float klow);
-	int ComputeGACLevelSet(float exponent,float kappa=5,float iso = .5);
-	int ComputeLaplacianLevelSet(float exponent,float kappa=5,float iso = .5);
+	int ComputeGACLevelSet(float expFactorRising);
 
 	int GenerateFeatureImage();
 	int GenerateSeedImage();
-	int CopyFrontToSeed();
 
 
 	//Front Image Stuff:
 	void DeallocateFrontObjs();
 	int GetFront(cvPolyData** front);
 	cvPolyData* GetFront();
+
 	cvStrPts* GetFrontImage();
 
 	//Get and Set Properties
@@ -149,9 +150,6 @@ public:
 
 	cvSetMacro(AdvectionScaling,double);
 	cvGetMacro(AdvectionScaling,double);
-
-	cvSetMacro(LaplacianScaling,double);
-	cvGetMacro(LaplacianScaling,double);
 
 	cvSetMacro(PropagationScaling,double);
 	cvGetMacro(PropagationScaling,double);
@@ -171,17 +169,11 @@ public:
 	cvSetMacro(Debug,bool);
 	cvGetMacro(Debug,bool);
 
-	cvSetMacro(UseNormalVectorCurvature,bool);
-	cvGetMacro(UseNormalVectorCurvature,bool);
+	cvSetMacro(UseInputImageAsFeature,bool);
+	cvGetMacro(UseInputImageAsFeature,bool);
 
-	cvSetMacro(UseMeanCurvature,bool);
-	cvGetMacro(UseMeanCurvature,bool);
-
-	cvSetMacro(UseMinimalCurvature,bool);
-	cvGetMacro(UseMinimalCurvature,bool);
-
-	cvSetMacro(BinarySeed,bool);
-	cvGetMacro(BinarySeed,bool);
+	cvSetMacro(UseInputImageDistance,bool);
+	cvGetMacro(UseInputImageDistance,bool);
 
 	//cvSetRepoObjMacro(InputImage,cvStrPts,vtkStructuredPoints);
 	int SetInputImage(cvStrPts *s);
@@ -209,9 +201,6 @@ public:
 	}
 #endif
 
-	template<typename TLevelSetFilterType>
-	void ShowDebug(typename TLevelSetFilterType::Pointer levelSetFilter);
-
 	void WriteFrontImages()
 	{
 		std::stringstream filenameBase;
@@ -221,8 +210,8 @@ public:
 
 	}
 
-	cvITKLevelSetBase();
-	virtual ~cvITKLevelSetBase()
+	cvITKLevelSet();
+	virtual ~cvITKLevelSet()
 	{
 		if ( m_cvInputImage != NULL ) {
 			delete m_cvInputImage;
@@ -242,8 +231,8 @@ protected:
 	//base class
 	void GenerateData(); //Might need a helper function
 	//I do not want the compiler to auto-generate these methods.
-	cvITKLevelSetBase(const cvITKLevelSetBase &); // purposely not implemented
-	void operator=(const cvITKLevelSetBase &); // purposely not implemented
+	cvITKLevelSet(const cvITKLevelSet &); // purposely not implemented
+	void operator=(const cvITKLevelSet &); // purposely not implemented
 	//void
 
 
@@ -255,8 +244,8 @@ private:
 	cvStructuredPoints* m_cvInputImage;
 	cvPolyData* m_cvSeed;
 
-	typename ITKInternalImageType::Pointer m_itkFeatureImage;
-	typename ITKInternalImageType::Pointer m_itkSeedImage;
+	ITKInternalImageType::Pointer m_itkFeatureImage;
+	ITKInternalImageType::Pointer m_itkSeedImage;
 
 	//output images
 	vtkSmartPointer<vtkPolyData> m_vtkFrontPolyData;
@@ -264,7 +253,7 @@ private:
 
 	//Helper and debug output images
 	vtkSmartPointer<vtkStructuredPoints> m_vtkFeatureImage;
-	typename ITKInternalImageType::Pointer m_itkFrontImage;
+	ITKInternalImageType::Pointer m_itkFrontImage;
 	cvStructuredPoints* m_cvSeedImage;
 
 	//Level Set Parameters
@@ -277,20 +266,16 @@ private:
 	double m_PropagationScaling;
 	double m_CurvatureScaling;
 	double m_AdvectionScaling;
-	double m_LaplacianScaling;
 
-	//curvature type
-	bool m_UseNormalVectorCurvature;
-	bool m_UseMeanCurvature;
-	bool m_UseMinimalCurvature;
-	bool m_BinarySeed;
+	bool m_UseInputImageAsFeature;
+	bool m_UseInputImageDistance;
 
 	//ITK Stuff
 	bool m_Debug;
 
 };
 
-#include "cvITKLevelSetBase.cxx"
+
 
 //throw e_; /* Explicit naming to work around Intel compiler bug.  */					\
